+++ app/Http/Controllers/TandaTerimav2Controller.php
@@ -46,10 +46,12 @@ class TandaTerimav2Controller extends Controller
             ->addColumn('action', function ($data) {
                
                 $action = '';
-                $action .= "<a href='" . route('tandaterima.print',$data->id_tandaterima) . "' class='btn btn-icon btn-success'><i class='fa fa-print'></i></a>&nbsp;"; 
-                $action .= "<a href='javascript:void(0)' class='btn btn-icon btn-primary' data-id='{$data->id_tandaterima}' onclick='showTandaTerima(this);'><i class='fa fa-edit'></i></a>&nbsp;";
-                $action .= "<a href='javascript:void(0)' class='btn btn-icon btn-danger'  data-id='{$data->id_tandaterima}' onclick='deleteTandaTerima(this);'><i class='fa fa-trash'></i></a>&nbsp;";
-
+                $id_berkas = $data->berkas->id_user;
+                if ($id_berkas ==  Auth::user()->id_user ||  Auth::user()->level_user == 'Admin' || Auth::user()->level_user == 'Superadmin') {
+                    $action .= "<a href='" . route('tandaterima.print',$data->id_tandaterima) . "' class='btn btn-icon btn-success'><i class='fa fa-print'></i></a>&nbsp;"; 
+                    $action .= "<a href='javascript:void(0)' class='btn btn-icon btn-primary' data-id='{$data->id_tandaterima}' onclick='showTandaTerima(this);'><i class='fa fa-edit'></i></a>&nbsp;";
+                    $action .= "<a href='javascript:void(0)' class='btn btn-icon btn-danger'  data-id='{$data->id_tandaterima}' onclick='deleteTandaTerima(this);'><i class='fa fa-trash'></i></a>&nbsp;";
+                }
                 return $action;
             })
             ->addColumn('tanggal', function ($data) {
